+++ src/bin/add.rs
@@ -4,12 +4,12 @@ use std::path::PathBuf;
 use backhand::filesystem::{FilesystemHeader, FilesystemReader, FilesystemWriter};
 use clap::Parser;
 
-/// Binary to add file to squashfs filesystem
+/// tool to add files to squashfs filesystems
 #[derive(Parser, Debug)]
-#[command(author, version, about, long_about = None)]
+#[command(author, version)]
 struct Args {
     /// Squashfs file
-    input: PathBuf,
+    filesystem: PathBuf,
 
     // Path of file to read, to write into squashfs
     file: PathBuf,
@@ -24,13 +24,15 @@ fn main() {
     let args = Args::parse();
 
     // read of squashfs
-    let file = File::open(args.input).unwrap();
+    let file = File::open(args.filesystem).unwrap();
     let filesystem = FilesystemReader::from_reader(file).unwrap();
     let mut filesystem = FilesystemWriter::same_as_existing(&filesystem).unwrap();
 
     // create new file
-    let bytes = std::fs::read(&args.file).unwrap();
-    filesystem.push_file(bytes, args.file_path, FilesystemHeader::default());
+    let mut new_file = File::open(&args.file).unwrap();
+    filesystem
+        .push_file(&mut new_file, args.file_path, FilesystemHeader::default())
+        .unwrap();
 
     // write new file
     let bytes = filesystem.to_bytes().unwrap();